Added missing dependency and updated assumptions in harness

+++ verification/cbmc/proofs/aws_cryptosdk_hdr_write/aws_cryptosdk_hdr_write_harness.c
@@ -20,20 +20,24 @@
 #include <proof_helpers/proof_allocators.h>
 #include <proof_helpers/utils.h>
 
-// A generator function as described in the comment in aws_cryptosdk_hash_elems_array_init_stub.c.
-// Also see line 33 of the Makefile.
+/**
+ * A generator function as described in the comment in aws_cryptosdk_hash_elems_array_init_stub.c.
+ * Also see line 33 of the Makefile.
+ */
 void array_list_item_generator(struct aws_array_list *elems) {
     assert(elems->item_size == sizeof(struct aws_hash_element));
     for (size_t index = 0; index < elems->length; ++index) {
         struct aws_hash_element *val = (struct aws_hash_element *)((uint8_t *)elems->data + (elems->item_size * index));
-        // Due to the checks in aws_cryptosdk_enc_ctx_size, no string can have a length > UINT16_MAX
+        /* Due to the checks in aws_cryptosdk_enc_ctx_size, no string can have a length > UINT16_MAX. */
         struct aws_string *key = ensure_string_is_allocated_nondet_length();
         __CPROVER_assume(aws_string_is_valid(key));
-        __CPROVER_assume(key->len <= UINT16_MAX);
+        /* Due to the cast to uint16, the entire size of the enc_ctx must be less than < UINT16_MAX
+         * This is a simple way to ensure this without a call to enc_ctx_size. */
+        __CPROVER_assume(key->len <= UINT8_MAX);
         val->key                 = key;
         struct aws_string *value = ensure_string_is_allocated_nondet_length();
         __CPROVER_assume(aws_string_is_valid(value));
-        __CPROVER_assume(value->len <= UINT16_MAX);
+        __CPROVER_assume(value->len <= UINT8_MAX);
         val->value = value;
     }
 }
